+++ Term2/Project1_Extended_Kalman_Filters/src/tools.cpp
@@ -50,15 +50,15 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
    */
   MatrixXd Hj(3,4);
   // recover state parameters
-  float px = x_state(0);
-  float py = x_state(1);
-  float vx = x_state(2);
-  float vy = x_state(3);
+  double px = x_state(0);
+  double py = x_state(1);
+  double vx = x_state(2);
+  double vy = x_state(3);
 
   // DONE: YOUR CODE HERE
 
   // check division by zero
-  static float eps = 0.0000001;
+  static double eps = 0.0000001;
   if( fabs(px)<eps && fabs(py)<eps )
   {
     Hj.setZero();
@@ -66,8 +66,8 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
   }
   
   // compute the Jacobian matrix
-  float px2 = px*px;
-  float py2 = py*py;
+  double px2 = px*px;
+  double py2 = py*py;
 
   Hj << px/(sqrt(px2+py2)), py/(sqrt(px2+py2)), 0, 0,
         -py/(px2+py2), px/(px2+py2), 0, 0,
